app/requirements.txt
fastapi==0.103.1
uvicorn[standard]==0.23.2
transformers==4.33.2
diffusers==0.21.1
torch==2.0.1
Pillow==10.0.0